+++ src/framework/data_collection/mod.rs
@@ -1,2 +1,3 @@
 #[allow(dead_code)]
 pub mod dataframe_conversions;
+pub mod records;
