+++ src/framework/data_collection/records.rs
@@ -1,0 +1,63 @@
+use crate::framework::plugins::base_plugin::WorldTimer;
+
+use crate::framework::py_modules::simulation_builder::{Name, RecordInitializer};
+use bevy::prelude::{Commands, Component, Entity, Query, Res, Resource, Transform};
+use polars::prelude::NamedFrom;
+use std::collections::HashMap;
+
+#[derive(Component, Default)]
+pub struct Record {
+    pub record_name: String,
+    pub record_output: bool,
+    pub dataframe: polars::frame::DataFrame,
+}
+
+#[derive(Resource)]
+pub struct DataframeStore(pub Box<HashMap<String, Box<polars::frame::DataFrame>>>);
+
+#[derive(Resource)]
+pub struct DataFrameSender(pub flume::Sender<Box<HashMap<String, Box<polars::frame::DataFrame>>>>);
+
+#[allow(unused_variables)]
+pub fn initialize_records(
+    mut commands: Commands,
+    query_entities: Query<(Entity, &Name, &RecordInitializer, &Transform)>,
+    world_timer: Res<WorldTimer>,
+) {
+    for (e, n, r_i, t) in query_entities.iter() {
+        // Get reference to position from the transform
+        let position = &t.translation;
+        // create new row to add to the dataframe
+        let new_row = polars::df!["Time" => [world_timer.timer.elapsed_secs()], "Position_X" => [position.x], "Position_Y" => [position.y], "Position_Z" => [position.z]].unwrap();
+
+        commands
+            .entity(e)
+            .insert(Record {
+                record_name: n.0.clone(),
+                record_output: true,
+                dataframe: new_row,
+            })
+            .remove::<RecordInitializer>();
+    }
+}
+
+/// system to query for all entities that have both Record and Transform components
+/// grabs data from the transform and records it to the record component
+pub fn update_records(
+    mut record_components: Query<(&mut Record, &Transform)>,
+    world_timer: Res<WorldTimer>,
+) {
+    // iterate over query results, destructuring the returned tuple
+    for (mut r, t) in record_components.iter_mut() {
+        // check if we should be recording data for this record component
+        if r.record_output {
+            // Get reference to position from the transform
+            let position = &t.translation;
+            // create new row to add to the dataframe
+            let new_row = polars::df!["Time" => [world_timer.timer.elapsed_secs()], "Position_X" => [position.x], "Position_Y" => [position.y], "Position_Z" => [position.z]].unwrap();
+
+            // Call vstack_mut function with the newly created row to append to dataframe
+            r.dataframe.vstack_mut(&new_row).unwrap();
+        }
+    }
+}
