+++ src/framework/plugins/mod.rs
@@ -1,2 +1,3 @@
 pub mod base_plugin;
+pub mod custom_plugins;
 pub mod plugin_group;
